requirements.txt
streamlit==1.32.2
pandas==1.5.3
numpy==1.26.4
google-cloud-bigquery==3.20.1
google-auth==2.29.0
DateTime==5.5
db-dtypes==1.2.0
python-dotenv==1.0.1